requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
scikit-learn
joblib
efficientnet_pytorch
openai
python-multipart
tensorflow 